requirements.txt
PyYAML==3.10
nose==1.3.0
wsgiref==0.1.2
